+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Features, Bug Fixes, API Breaking, Deprecated, Infrastructure, Template Updates
 ### Breaking Changes
 
 - The method `ProduceBlock` has been renamed to `ProduceBlocks(numberOfBlocks=1)` and now accepts optional parameter that allows user to specify the number of blocks that will be produced.
+- `appendSignMultisigTransaction` now expects `Encoded MultiSignature Transaction` as first argument and returns an object containing a blob key encoded in base64.
 
 #### TEALv8